+++ test-py/test_path_changes.py
@@ -4,7 +4,7 @@ import pathlib
 import pytest
 
 from fontra.core.changes import applyChange
-from fontra.core.packedpath import PackedPath
+from fontra.core.path import PackedPath
 
 testDataPath = (
     pathlib.Path(__file__).parent.parent / "test-common" / "path-change-test-data.json"
