+++ test-py/test_path_io.py
@@ -3,7 +3,7 @@ from dataclasses import asdict
 import pytest
 
 from fontra.core.classes import from_dict
-from fontra.core.packedpath import PackedPath, PackedPathPointPen
+from fontra.core.path import PackedPath, PackedPathPointPen
 
 pathTestData = [
     {
